+++ lib/raavailability.dart
@@ -5,8 +5,8 @@ import 'package:flutter/services.dart';
 class Raavailability {
   static const MethodChannel _channel = const MethodChannel('raavailability');
 
-  static Future<String> get isSupported async {
-    final String supported = await _channel.invokeMethod('isSupported');
+  static Future<bool> get isSupported async {
+    final bool supported = await _channel.invokeMethod('isSupported');
     return supported;
   }
 }