+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.0.1
-Implementing the isSupported functionr for android.
+Implementing the isSupported function for android.
 ## 0.0.2
 Changing plugin name, running flutter format and add a homepage url.
 ## 0.0.3
@@ -10,3 +10,5 @@ Improoving Example file.
 Changing dependencies place.
 ## 0.0.6
 Changing google ar core version.
+## 0.1.0
+Using requestInstall instead of isSupported